GVT-1665 Move geometry plan start address calculation to GeometryPlanLayout construction time
